+++ HISTORY.rst
@@ -2,6 +2,12 @@
 History
 =======
 
+0.3.2 (2017-09-18)
+------------------
+
+* All plots except for HivePlot are implemented.
+* Implemented auto-colorbar for plots that have continuous node colors.
+
 0.1.0 (2016-07-15)
 ------------------